+++ index.html
@@ -116,7 +116,8 @@
                 <!-- Добавление хэш-тегов и комментария к изображению -->
                 <fieldset class="img-upload__text text">
                   <div class="img-upload__field-wrapper">
-                    <input type="text" class="text__hashtags" name="hashtags" placeholder="#ХэшТег">
+                    <input type="text" class="text__hashtags" name="hashtags" placeholder="#ХэшТег"
+                    data-pristine-maxlength-message = "Хештегов должно быть пять! ">
                   </div>
                   <div class="img-upload__field-wrapper">
                     <textarea class="text__description" name="description" placeholder="Ваш комментарий..." data-pristine-maxlength = "140"
@@ -156,7 +157,7 @@
 
 
           <!-- Комментарии к изображению-->
-          <div class="social__comment-count hidden">5</span> из <span class="comments-count">125</span> комментариев</div>
+          <div class="social__comment-count"><span class="shown">5</span> из <span class="comments-count">125</span> комментариев</div>
           <ul class="social__comments">
             <li class="social__comment">
               <img class="social__picture" src="img/avatar-4.svg" alt="Аватар комментатора фотографии" width="35" height="35">
